refactor: move 3d math to its own module
