+++ lvgl.h
@@ -27,6 +27,7 @@ extern "C" {
 #include "src/misc/lv_log.h"
 #include "src/misc/lv_timer.h"
 #include "src/misc/lv_math.h"
+#include "src/misc/lv_3dmath.h"
 #include "src/misc/lv_array.h"
 #include "src/misc/lv_async.h"
 #include "src/misc/lv_anim_timeline.h"
