+++ src/misc/lv_3dmath.c
@@ -1,0 +1,47 @@
+/**
+ * @file lv_3dmath.c
+ *
+ */
+
+/*********************
+ *      INCLUDES
+ *********************/
+
+#include "lv_3dmath.h"
+
+/*********************
+ *      DEFINES
+ *********************/
+
+/**********************
+ *      TYPEDEFS
+ **********************/
+
+/**********************
+ *  STATIC PROTOTYPES
+ **********************/
+
+/**********************
+ *  STATIC VARIABLES
+ **********************/
+
+/**********************
+ *      MACROS
+ **********************/
+
+/**********************
+ *   GLOBAL FUNCTIONS
+ **********************/
+
+lv_3dplane_t lv_get_ground_plane(float elevation)
+{
+    return (lv_3dplane_t) {
+        .origin = {0.f, elevation, 0.f},
+        .direction = {0.f, 1.f, 0.f}
+    };
+}
+
+
+/**********************
+ *   STATIC FUNCTIONS
+ **********************/
