+++ src/misc/lv_math.h
@@ -41,26 +41,6 @@ typedef struct {
     uint16_t f;
 } lv_sqrt_res_t;
 
-typedef struct {
-    float x;
-    float y;
-    float z;
-} lv_3dpoint_t;
-
-typedef struct {
-    float x;
-    float y;
-    float z;
-    float w;
-} lv_quaternion_t;
-
-typedef struct {
-    lv_3dpoint_t origin;
-    lv_3dpoint_t direction;
-} lv_3dplane_t;
-
-typedef lv_3dplane_t lv_3dray_t;
-
 /**********************
  * GLOBAL PROTOTYPES
  **********************/
@@ -165,14 +145,6 @@ void lv_rand_set_seed(uint32_t seed);
  * @return return the random number. min <= return_value <= max
  */
 uint32_t lv_rand(uint32_t min, uint32_t max);
-
-
-/**
- * Get a plane that faces upward, centered at a given height
- * @param elevation elevation of the ground plane, in world units. this is usually zero
- * @return ground plane
- */
-lv_3dplane_t lv_get_ground_plane(float elevation);
 
 /**********************
  *      MACROS
